Tell Windows users to build with '--disable-shared'
As detailed in GitHub issue #1, attempting to build ppamltracer as a
shared library on Cygwin fails.  I believed commit bfc92ba (‘autotools:
Don’t build DLLs on Cygwin by default’) fixed this; it does, but it also
disables shared library building on non-Cygwin systems.

We need to do a release, so this bug is getting deferred until later;
for the time being, the documented workaround is to pass
'--disable-shared' to 'configure'.