Link against librt if necessary
For a while, we were very confused why ppamltracer would work happily on
some Debian systems but produce dynamic linker errors on other, similar
ones.  It turns out the difference was glibc versions – if you have
glibc <2.17, you need to specify '-lrt' for one of the functions
ppamltracer uses.  This commit adds an Autoconf macro that adds that
flag if necessary; it should improve portability in general as well.